requirements.txt
streamlit
python-docx
openai
requests